+++ src/main/java/com/epam/reportportal/service/step/DefaultStepReporter.java
@@ -220,9 +220,14 @@ public class DefaultStepReporter implements StepReporter {
 	}
 
 	@Override
-	public void finishNestedStep() {
-		FinishTestItemRQ finishStepRequest = buildFinishTestItemRequest(ItemStatus.PASSED);
+	public void finishNestedStep(@Nonnull ItemStatus status) {
+		FinishTestItemRQ finishStepRequest = buildFinishTestItemRequest(status);
 		finishNestedStep(finishStepRequest);
+	}
+
+	@Override
+	public void finishNestedStep() {
+		finishNestedStep(ItemStatus.PASSED);
 	}
 
 	@Override
@@ -253,7 +258,6 @@ public class DefaultStepReporter implements StepReporter {
 			finishNestedStep(buildFinishTestItemRequest(ItemStatus.PASSED));
 			return result;
 		} catch (RuntimeException | Error e) {
-			ReportPortal.emitLog(itemUuid -> buildSaveLogRequest(itemUuid, e));
 			finishNestedStep(buildFinishTestItemRequest(ItemStatus.FAILED));
 			throw e;
 		}